+++ src/model/PayPalLegacySettings.js
@@ -1,0 +1,108 @@
+/**
+ * DocuSign REST API
+ * The DocuSign REST API provides you with a powerful, convenient, and simple Web services API for interacting with DocuSign.
+ *
+ * OpenAPI spec version: v2.1
+ * Contact: devcenter@docusign.com
+ *
+ * NOTE: This class is auto generated. Do not edit the class manually and submit a new issue instead.
+ *
+ */
+
+(function(root, factory) {
+  if (typeof define === 'function' && define.amd) {
+    // AMD. Register as an anonymous module.
+    define(['ApiClient'], factory);
+  } else if (typeof module === 'object' && module.exports) {
+    // CommonJS-like environments that support module.exports, like Node.
+    module.exports = factory(require('../ApiClient'));
+  } else {
+    // Browser globals (root is window)
+    if (!root.Docusign) {
+      root.Docusign = {};
+    }
+    root.Docusign.PayPalLegacySettings = factory(root.Docusign.ApiClient);
+  }
+}(this, function(ApiClient) {
+  'use strict';
+
+
+  /**
+   * The PayPalLegacySettings model module.
+   * @module model/PayPalLegacySettings
+   * @version 3.0.0
+   */
+
+  /**
+   * Constructs a new <code>PayPalLegacySettings</code>.
+   * @alias module:model/PayPalLegacySettings
+   * @class
+   */
+  var exports = function() {
+    var _this = this;
+
+
+  };
+
+  /**
+   * Constructs a <code>PayPalLegacySettings</code> from a plain JavaScript object, optionally creating a new instance.
+   * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
+   * @param {Object} data The plain JavaScript object bearing properties of interest.
+   * @param {module:model/PayPalLegacySettings} obj Optional instance to populate.
+   * @return {module:model/PayPalLegacySettings} The populated <code>PayPalLegacySettings</code> instance.
+   */
+  exports.constructFromObject = function(data, obj) {
+    if (data) {
+      obj = obj || new exports();
+
+      if (data.hasOwnProperty('currency')) {
+        obj['currency'] = ApiClient.convertToType(data['currency'], 'String');
+      }
+      if (data.hasOwnProperty('partner')) {
+        obj['partner'] = ApiClient.convertToType(data['partner'], 'String');
+      }
+      if (data.hasOwnProperty('password')) {
+        obj['password'] = ApiClient.convertToType(data['password'], 'String');
+      }
+      if (data.hasOwnProperty('userName')) {
+        obj['userName'] = ApiClient.convertToType(data['userName'], 'String');
+      }
+      if (data.hasOwnProperty('vendor')) {
+        obj['vendor'] = ApiClient.convertToType(data['vendor'], 'String');
+      }
+    }
+    return obj;
+  }
+
+  /**
+   * 
+   * @member {String} currency
+   */
+  exports.prototype['currency'] = undefined;
+  /**
+   * 
+   * @member {String} partner
+   */
+  exports.prototype['partner'] = undefined;
+  /**
+   * 
+   * @member {String} password
+   */
+  exports.prototype['password'] = undefined;
+  /**
+   * 
+   * @member {String} userName
+   */
+  exports.prototype['userName'] = undefined;
+  /**
+   * 
+   * @member {String} vendor
+   */
+  exports.prototype['vendor'] = undefined;
+
+
+
+  return exports;
+}));
+
+